+++ src/ssh/azext_ssh/custom.py
@@ -21,6 +21,7 @@ from . import ssh_utils
 
 def ssh_vm(cmd, resource_group_name=None, vm_name=None, ssh_ip=None, public_key_file=None,
            private_key_file=None, use_private_ip=False, port=None, ssh_args=None):
+    _assert_args(resource_group_name, vm_name, ssh_ip)
     credentials_folder = None
     op_call = functools.partial(ssh_utils.start_ssh_connection, port, ssh_args)
     _do_ssh_op(cmd, resource_group_name, vm_name, ssh_ip,
@@ -29,14 +30,19 @@ def ssh_vm(cmd, resource_group_name=None, vm_name=None, ssh_ip=None, public_key_
 
 def ssh_config(cmd, config_path, resource_group_name=None, vm_name=None, ssh_ip=None,
                public_key_file=None, private_key_file=None, overwrite=False, use_private_ip=False, credentials_folder=None):
+    _assert_args(resource_group_name, vm_name, ssh_ip)
     # If user provides their own key pair, certificate will be written in the same folder as public key.
     if (public_key_file or private_key_file) and credentials_folder:
-        raise azclierror.InvalidArgumentValueError("If providing --public-key-file/-p or --private-key-file/-i, --keys-destination-folder should not be used")
+        raise azclierror.ArgumentUsageError("--keys-destination-folder can't be used in conjunction with --public-key-file/-p or --private-key-file/-i.")
     op_call = functools.partial(ssh_utils.write_ssh_config, config_path, resource_group_name, vm_name, overwrite)
     # Default credential location
     if not credentials_folder:
         config_folder = os.path.dirname(config_path)
-        folder_name = resource_group_name + "-" + vm_name
+        if not os.path.isdir(config_folder):
+            raise azclierror.InvalidArgumentValueError(f"Config file destination folder {config_folder} does not exist.")
+        folder_name = ssh_ip
+        if resource_group_name and vm_name:
+            folder_name = resource_group_name + "-" + vm_name
         credentials_folder = os.path.join(config_folder, os.path.join("az_ssh_config", folder_name))
 
     _do_ssh_op(cmd, resource_group_name, vm_name, ssh_ip, public_key_file, private_key_file, use_private_ip, credentials_folder, op_call)
@@ -45,12 +51,11 @@ def ssh_config(cmd, config_path, resource_group_name=None, vm_name=None, ssh_ip=
 def ssh_cert(cmd, cert_path=None, public_key_file=None):
     if not cert_path and not public_key_file:
         raise azclierror.RequiredArgumentMissingError("--file or --public-key-file must be provided.")
-    # If user doesn't provide a public key, save key to the same folder as --file
+    # If user doesn't provide a public key, save generated key pair to the same folder as --file
     keys_folder = None
     if not public_key_file:
         keys_folder = os.path.dirname(cert_path)
-        logger.warning("No public key provided. A new key pair will be created in the same directory as the certificate (%s). "
-                       "Please delete keys once the certificate is no longer being used",
+        logger.warning("The generated SSH keys are stored at %s. Please delete SSH keys when the certificate is no longer being used.",
                        keys_folder)
     public_key_file, _, _ = _check_or_create_public_private_files(public_key_file, None, keys_folder)
     cert_file, _ = _get_and_write_certificate(cmd, public_key_file, cert_path)
@@ -58,7 +63,6 @@ def ssh_cert(cmd, cert_path=None, public_key_file=None):
 
 
 def _do_ssh_op(cmd, resource_group, vm_name, ssh_ip, public_key_file, private_key_file, use_private_ip, credentials_folder, op_call):
-    _assert_args(resource_group, vm_name, ssh_ip)
     # Get ssh_ip before getting public key to avoid getting "ResourceNotFound" exception after creating the keys
     ssh_ip = ssh_ip or ip_utils.get_ssh_ip(cmd, resource_group, vm_name, use_private_ip)
 
@@ -148,6 +152,7 @@ def _check_or_create_public_private_files(public_key_file, private_key_file, cre
     # If nothing is passed in create a directory with a ephemeral keypair
     if not public_key_file and not private_key_file:
         # We only want to delete the keys if the user hasn't providede their own keys
+        # Only ssh vm deletes generated keys. 
         delete_keys = True
         if not credentials_folder:
             # az ssh vm: Create keys on temp folder and delete folder once connection succeeds/fails.